+++ src/Raven.CodeAnalysis/Macros/MacroKind.cs
@@ -3,4 +3,5 @@ namespace Raven.CodeAnalysis.Macros;
 public enum MacroKind
 {
     AttachedDeclaration = 0,
+    FreestandingExpression = 1,
 }